refactor: move linked roles to core package
