+++ .changeset/itchy-spiders-grin.md
@@ -1,0 +1,5 @@
+---
+"@buape/carbon": minor
+---
+
+refactor: move linked roles to core package
