+++ apps/rocko/package.json
@@ -10,7 +10,6 @@
 	},
 	"dependencies": {
 		"@buape/carbon": "workspace:*",
-		"@buape/carbon-linked-roles": "workspace:*",
 		"@buape/carbon-nodejs": "workspace:*",
 		"@types/node": "22.5.5"
 	},
