+++ packages/carbon/package.json
@@ -4,6 +4,16 @@
 	"type": "module",
 	"main": "./dist/src/index.js",
 	"repository": "github:buape/carbon",
+	"exports": {
+		".": {
+			"types": "./dist/src/index.d.ts",
+			"import": "./dist/src/index.js"
+		},
+		"./*": {
+			"types": "./dist/src/plugins/*/index.d.ts",
+			"import": "./dist/src/plugins/*/index.js"
+		}
+	},
 	"scripts": {
 		"build": "tsc",
 		"dev": "tsc -w",
