+++ packages/carbon/typedoc.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://typedoc.org/schema.json",
 	"extends": ["../../typedoc.json"],
-	"entryPoints": ["src/index.ts"],
+	"entryPoints": ["src/index.ts", "src/plugins/*/index.ts"],
 	"name": "@buape/carbon",
 	"out": "docs",
 	"excludeExternals": true,
