--- packages/linked-roles/CHANGELOG.md
@@ -1,104 +1,0 @@
-# @buape/carbon-linked-roles
-
-## 0.2.2
-
-### Patch Changes
-
-- Updated dependencies [63a6d07]
-  - @buape/carbon@0.5.0
-
-## 0.2.1
-
-### Patch Changes
-
-- 6b73575: chore: readme cleanups
-- Updated dependencies [6b73575]
-  - @buape/carbon@0.4.2
-
-## 0.2.0
-
-### Minor Changes
-
-- 07b0817: fix: make side packages use peer dependencies
-
-### Patch Changes
-
-- Updated dependencies [558b73c]
-- Updated dependencies [28f252f]
-- Updated dependencies [50e360e]
-- Updated dependencies [4475a84]
-- Updated dependencies [29f8493]
-- Updated dependencies [5bc4b84]
-- Updated dependencies [6239590]
-- Updated dependencies [0a22fa8]
-- Updated dependencies [9456bf0]
-  - @buape/carbon@0.4.0
-
-## 0.1.7
-
-### Patch Changes
-
-- Updated dependencies [dcf5b44]
-- Updated dependencies [c8e7c1c]
-  - @buape/carbon@0.3.2
-
-## 0.1.6
-
-### Patch Changes
-
-- Updated dependencies [c80ce0f]
-  - @buape/carbon@0.3.1
-
-## 0.1.5
-
-### Patch Changes
-
-- Updated dependencies [f8b608f]
-- Updated dependencies [50c8200]
-- Updated dependencies [a15bf55]
-- Updated dependencies [51d84c3]
-  - @buape/carbon@0.3.0
-
-## 0.1.4
-
-### Patch Changes
-
-- Updated dependencies [9e93027]
-- Updated dependencies [ad9666b]
-  - @buape/carbon@0.2.0
-
-## 0.1.3
-
-### Patch Changes
-
-- Updated dependencies
-  - @buape/carbon@0.1.3
-
-## 0.1.2
-
-### Patch Changes
-
-- Updated dependencies [2bdc7f3]
-- Updated dependencies [e79c435]
-- Updated dependencies [917416a]
-  - @buape/carbon@0.1.2
-
-## 0.1.1
-
-### Patch Changes
-
-- Updated dependencies [8438b8d]
-  - @buape/carbon@0.1.1
-
-## 0.1.0
-
-### Minor Changes
-
-- 1a00131: Initial beta!
-
-  Featuring interactions, commands, messages, subcommands, options, linked roles, and more!
-
-### Patch Changes
-
-- Updated dependencies [1a00131]
-  - @buape/carbon@0.1.0
