--- packages/linked-roles/README.md
@@ -1,13 +1,0 @@
-# @buape/carbon-linked-roles
-
-<div align="center">
-<a href="https://go.buape.com/carbon"><img src="https://cdn.buape.com/CarbonWordmark.png" alt="Carbon Wordmark"></a>
-
-<img alt="Discord" src="https://img.shields.io/discord/1280628625904894072?style=for-the-badge">
-<img alt="NPM Version" src="https://img.shields.io/npm/v/@buape/carbon?style=for-the-badge">
-<img alt="NPM Downloads" src="https://img.shields.io/npm/dm/@buape/carbon?style=for-the-badge">
-</div>
-
-This package is part of [the Carbon framework](https://github.com/buape/carbon), check it out for more details.
-
-You can find the documentation for this package [here](https://carbon.buape.com/carbon/linked-roles).
