--- packages/linked-roles/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "@buape/carbon-linked-roles",
-	"version": "0.2.2",
-	"type": "module",
-	"main": "./dist/src/index.js",
-	"repository": "github:buape/carbon",
-	"scripts": {
-		"build": "tsc",
-		"dev": "tsc -w",
-		"docs": "typedoc"
-	},
-	"license": "MIT",
-	"peerDependencies": {
-		"@buape/carbon": "workspace:^0.5.0"
-	},
-	"devDependencies": {
-		"@buape/carbon": "workspace:*"
-	},
-	"files": [
-		"dist",
-		"LICENSE"
-	]
-}
