--- packages/linked-roles/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"include": [
-		"src/**/*.ts",
-		"src/*.ts"
-	],
-	"compilerOptions": {
-		"outDir": "dist",
-		"rootDir": "."
-	}
-}
