--- packages/linked-roles/typedoc.json
@@ -1,7 +1,0 @@
-{
-	"$schema": "https://typedoc.org/schema.json",
-	"extends": ["../../typedoc.json"],
-	"entryPoints": ["src/index.ts"],
-	"name": "@buape/carbon-linked-roles",
-	"out": "docs"
-}
