+++ website/content/carbon/meta.json
@@ -8,6 +8,7 @@
 		"...",
 		"--- Usage ---",
 		"classes",
+		"plugins",
 		"helpful-guides",
 		"even-more",
 		"--- API Reference ---",
