+++ website/content/carbon/plugins/index.mdx
@@ -1,0 +1,6 @@
+---
+title: Plugins
+index: true
+icon: Cable
+---
+
