+++ website/content/carbon/plugins/meta.json
@@ -1,0 +1,1 @@
+{ "title": "Plugins" }
