--- website/content/linked-roles/meta.json
@@ -1,4 +1,0 @@
-{
-	"root": true,
-	"pages": ["--- Usage ---", "index", "--- API Reference ---", "...api"]
-}
